+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v2.1.0
+
+- Added new updated job search response
+- Added tests for ResponseFactory
+
+# v2.0.0
+
+- Breaking Change Version
+- the authenticate() function must be called to activate authentication
+
 # v1.3.0
 
 - Adding a /search/jobs response factory helper